requirements.txt
streamlit
openai
requests
num2words
pandas
numpy
opencv-python
python-dotenv
nextcord
flask